# backend/config/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Alice Voice Backend API",
    "description": "REST and socket contract owned by the backend. Frontend copies DTOs from this file.",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787",
      "description": "Local daemon"
    }
  ],
  "paths": {
...
    "/api/v1/librarian/facts/{id}": {
      "delete": {
        "tags": [
          "librarian"
        ],
        "summary": "Reply of `DELETE /api/v1/librarian/facts/{id}`.",
        "description": "The fact is retired rather than deleted: it keeps its place in the\nhistory and stops being true now. The reply carries the concept it\nbelongs to, so the editor reads the new state without a second call.",
        "operationId": "retire_fact_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifier of the fact",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The concept the fact belongs to",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryNodeDto"
                }
              }
            }
          },
          "404": {
            "description": "The fact does not exist"
          }
        }
      }
    },
...
    "/api/v1/librarian/memories": {
      "get": {
        "tags": [
          "librarian"
        ],
        "summary": "Reply of `GET /api/v1/librarian/memories`.",
        "operationId": "list_memories_handler",
        "parameters": [
          {
            "name": "text",
            "in": "query",
            "description": "Words to search the memory for",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Largest number of nodes to return",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Memory nodes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryQueryDto"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "librarian"
        ],
        "summary": "Reply of `POST /api/v1/librarian/memories`.",
        "operationId": "write_memory_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryWriteRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The stored node",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryNodeDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/api/v1/librarian/memories/{id}": {
      "delete": {
        "tags": [
          "librarian"
        ],
        "summary": "Reply of `DELETE /api/v1/librarian/memories/{id}`.",
        "description": "The whole concept goes, facts included. A reader who wants to keep the\nhistory retires the one fact instead.",
        "operationId": "delete_memory_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifier of the concept",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The concept is deleted"
          },
          "404": {
            "description": "The concept does not exist"
          }
        }
      }
    },
    "/api/v1/librarian/memories/{id}/merge": {
      "post": {
        "tags": [
          "librarian"
        ],
        "summary": "Reply of `POST /api/v1/librarian/memories/{id}/merge`.",
        "description": "The concept in the path moves into the concept the body names, so a\nreader settles the two concepts the memory holds for one thing without\ndeleting either of them by hand. The reply is the concept that\nabsorbed the other, so the editor reads its new state without a second\ncall.",
        "operationId": "merge_memory_handler",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Identifier of the concept that moves",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryMergeRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The concept that absorbed the other",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryNodeDto"
                }
              }
            }
          },
          "400": {
            "description": "The merge names one concept twice"
          },
          "404": {
            "description": "One of the concepts does not exist"
          }
        }
      }
    },
    "/api/v1/librarian/query": {
      "post": {
        "tags": [
          "librarian"
        ],
        "summary": "Reply of `POST /api/v1/librarian/query`.",
        "operationId": "query_memory_handler",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoryQueryRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Memory nodes",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryQueryDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MemoryFactDto": {
        "type": "object",
        "description": "One dated relation of one memory node.",
        "required": [
          "id",
          "relation",
          "value",
          "current",
          "confidence",
          "importance",
          "confirmations",
          "lastConfirmedAt",
          "validAt"
        ],
        "properties": {
          "confidence": {
            "type": "number",
            "format": "float",
            "description": "How sure the reader was of the fact, from zero to one."
          },
          "confirmations": {
            "type": "integer",
            "format": "int32",
            "description": "Number of later turns that taught the same fact again."
          },
          "current": {
            "type": "boolean",
            "description": "Whether the fact is true now."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Stable identifier of the fact."
          },
          "importance": {
            "type": "number",
            "format": "float",
            "description": "How much the fact is worth, from zero to one."
          },
          "invalidAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time",
            "description": "Time the fact stopped being true, or null while it is current."
          },
          "lastConfirmedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Time the memory last saw the fact, in ISO 8601."
          },
          "relation": {
            "type": "string",
            "description": "The relation the fact names, for example `lives_in`."
          },
          "validAt": {
            "type": "string",
            "format": "date-time",
            "description": "Time the fact became true, in ISO 8601."
          },
          "value": {
            "type": "string",
            "description": "The value the fact carries."
          }
        }
      },
      "MemoryFactWriteDto": {
        "type": "object",
        "description": "One relation a write gives to a node.",
        "required": [
          "relation",
          "value"
        ],
        "properties": {
          "relation": {
            "type": "string",
            "description": "The relation the fact names."
          },
          "value": {
            "type": "string",
            "description": "The value the fact carries."
          }
        }
      },
      "MemoryMergeRequestDto": {
        "type": "object",
        "description": "Body of a memory merge.\n\nThe concept the request names keeps its key and its facts, and the\nconcept in the path moves into it: its facts move as they stand, its\nkey becomes a name of the target, and the empty concept goes.",
        "required": [
          "into"
        ],
        "properties": {
          "into": {
            "type": "string",
            "format": "uuid",
            "description": "Identifier of the concept that absorbs the other."
          }
        }
      },
      "MemoryNodeDto": {
        "type": "object",
        "description": "One memory node with its facts.",
        "required": [
          "id",
          "key",
          "title",
          "body",
          "facts",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "body": {
            "type": "string",
            "description": "The body of the concept in plain text."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Time the daemon stored the node, in ISO 8601."
          },
          "facts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemoryFactDto"
            },
            "description": "The relations of the node, current first."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Stable identifier of the node."
          },
          "key": {
            "type": "string",
            "description": "The key that names the node, for example `user`."
          },
          "title": {
            "type": "string",
            "description": "The name the memory shows."
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Time the daemon last changed the node, in ISO 8601."
          }
        }
      },
      "MemoryQueryDto": {
        "type": "object",
        "description": "Reply of a memory query.",
        "required": [
          "nodes"
        ],
        "properties": {
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemoryNodeDto"
            },
            "description": "The nodes the query matched, best first."
          }
        }
      },
      "MemoryQueryRequestDto": {
        "type": "object",
        "description": "Body of the memory query.",
        "required": [
          "text"
        ],
        "properties": {
          "limit": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Largest number of nodes to return.",
            "minimum": 0
          },
          "text": {
            "type": "string",
            "description": "The words to search the memory for. Empty returns the recent nodes."
          }
        }
      },
...
      "MemoryWriteRequestDto": {
        "type": "object",
        "description": "Body of a memory write.",
        "required": [
          "key",
          "title",
          "body",
          "facts"
        ],
        "properties": {
          "body": {
            "type": "string",
            "description": "The body of the concept."
          },
          "facts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MemoryFactWriteDto"
            },
            "description": "The relations to write. A relation with a new value closes the\nolder value rather than deleting it."
          },
          "key": {
            "type": "string",
            "description": "The key that names the node. An existing key is enriched."
          },
          "title": {
            "type": "string",
            "description": "The name the memory shows."
          }
        }
      },
...
    }
  }
}